Extract Core from shared Fn/Imp things
