+++ purescript.cabal
@@ -57,13 +57,14 @@ library
                      Language.PureScript.CodeGen.JS.Optimizer.TCO
                      Language.PureScript.CodeGen.JS.Optimizer.Unused
                      Language.PureScript.Constants
+                     Language.PureScript.Core
+                     Language.PureScript.Core.Ann
+                     Language.PureScript.Core.Literals
+                     Language.PureScript.Core.Meta
                      Language.PureScript.CoreFn
-                     Language.PureScript.CoreFn.Ann
                      Language.PureScript.CoreFn.Binders
                      Language.PureScript.CoreFn.Desugar
                      Language.PureScript.CoreFn.Expr
-                     Language.PureScript.CoreFn.Literals
-                     Language.PureScript.CoreFn.Meta
                      Language.PureScript.CoreFn.Module
                      Language.PureScript.CoreFn.Traversals
                      Language.PureScript.CoreImp
