+++ src/Language/PureScript/CoreFn.hs
@@ -16,11 +16,8 @@ module Language.PureScript.CoreFn (
   module C
 ) where
 
-import Language.PureScript.CoreFn.Ann as C
 import Language.PureScript.CoreFn.Binders as C
 import Language.PureScript.CoreFn.Desugar as C
 import Language.PureScript.CoreFn.Expr as C
-import Language.PureScript.CoreFn.Literals as C
-import Language.PureScript.CoreFn.Meta as C
 import Language.PureScript.CoreFn.Module as C
 import Language.PureScript.CoreFn.Traversals as C
